add uuid column in coin history table.
